Fix check to ensure we have valid mktPrice
The current logic throws an error if the first observation of the
mktPrice object is NA, even if that observation isn't being used.

Check for NULL first, since it will almost always be FALSE. Then check
if the length of mktPrice==1 and that one value is NA.

Thanks to Kyle Balkissoon for the report and reproducible example.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/blotter@1706 edb9625f-4e0d-4859-8d74-9fd3b1da38cb