Fix signal analysis with only one distribution
The signal analysis code would only run the first distribution value if
there is only one distribution, because it subset a data.frame with
only a value for 'i'. If there's only one distrubtion, there's only one
column in the data.frame, so subsetting by 'i' returns a vector.

Add a check in install.param.combo to throw an error if it is passed a
vector instead of an object with a dim attribute.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/blotter@1729 edb9625f-4e0d-4859-8d74-9fd3b1da38cb