Fix unit tests
Some function names have changed to include a leading dot, and they are
no longer exported, so we need to access them with the ::: operator.
Thanks to cloudcello for the report and patch (#6284).

Make transaction fees negative when calling addTxn. Change portfolio
and account names to avoid clashing with anything a user might have
created.

Clean up objects and environments when tests exit (instead of deleting
everything). I couldn't figure out how to do it with .setUp/.tearDown,
so I used on.exit instead. The issue with .setUp/.tearDown was likely
due to scoping issues, and I didn't want to assign to the .GlobalEnv.

Remove test.calcRealizedPL because calcRealizedPL was removed in r411.
Add test.calcPosAvgCost_C to test C version of .calcPosAvgCost.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/blotter@1732 edb9625f-4e0d-4859-8d74-9fd3b1da38cb